requirements.txt
Pillow
requests
torch
ffmpeg
git+https://github.com/openai/whisper.git
moviepy
gradio
transformers
opencv-python
accelerate